# Faster vec_extract: Testing requested