@AethanFoot 2021-04-12
Extra question, should this function return a VecDeque as they are more efficient with reordering (then convert back to a vector before appending).