@lex148 2021-04-12
I agree with you switching to drain_filter is the best option when it is out of nightly. I will give this one a good testing tomorrow...